Change entity id from unsigned to size_t

+++ include/sprites.hpp
@@ -65,7 +65,7 @@ using Sprites = seb_engine::Sprites<MAX_ENTITIES, SpriteBase, SpriteHead, Sprite
 
 namespace sprites
 {
-auto lookup_set_movement_sprites(Sprites& sprites, unsigned id, Entity entity, raylib::Vector2 vel) -> void;
+auto lookup_set_movement_sprites(Sprites& sprites, size_t id, Entity entity, raylib::Vector2 vel) -> void;
 [[nodiscard]] auto alternate_frame_y_offset(SpriteLegs legs) -> float;
 [[nodiscard]] auto flipped_x_offset(raylib::Vector2 sprite_size) -> float;
 } // namespace sprites
